Add files via upload
This contains the most recent changes which include:
- item list displayed
- item fetching from database in real time

Note: there is a bug in displaying item description. When a user clicks on an item in the list it should redirect the user to a new page with the item description. However, it stays in a constant loop checking for the Item to be loaded.